refactor: rename function
